[libcrl] Rewards are doubles to address numerical issues in ALP (specifically: backprojections).

+++ libcrl/include/crl/types.hpp
@@ -28,7 +28,7 @@
 
 namespace crl {
 
-typedef float Reward;
+typedef double Reward;
 typedef double Probability;
 typedef short Factor;  ///< The value type of a (discrete) Factor in the graph
 typedef uint64_t Size;
